Migrate `bitcore-lib-xpi` to ECMAScript, TypeScript (#38)
This large commit tackles and important milestone: migrating from the
legacy `bitcore-lib-xpi` NPM package that has been in use for several
years.

Much of the key functionality was migrated and tested, including
HDPrivateKey, mainnet/testnet/regtest addresses, P2SH support, and more.

IMPORTANT: This is only a preliminary migration and SOME THINGS MAY NOT
WORK

Also, the `@abcpros/bitcore-mnemonic` package was migrated to be
included directly with this new Bitcore implementation.