Fixup the Script ASM, OP codes (#50)
This commit adds the `asm` property to the `Output` class and also adds
the `changeAsm` property to the `Transaction` class.

Also, the ASM generation function in the `Script` class was updated to
use the existing `Opcode.map` in the `Opcode` class. Also,
`OP_SCRIPTTYPE` replaced the `OP_VER` OP code.