Add missing depends, remove `uint8arrays` (#54)
Primarily, `chronik-client` was missing from the `package.json` after
being implemented in a previous commit. This commit adds that
dependency.

We also remove the unnecessary `uint8arrays` dependency and adjust the
modules to use native NodeJS `Buffer` objects. This *could* cause an
issue for browsers later; will keep an eye on it.