Add NAT-T packages, configuration (#55)
This commit adds new dependencies required for setting up NAT-T relays
and client-side configuration.

There are new configuration options for the core P2P/DHT modules that
allow nodes to run as a relay server. This is ideally set up on the
bootstrap nodes, but as of now is not restricted to any node on the
network.

The P2PCoordinator class is also updated to add the additional
transports and helper modules required to upgrade relay connections to
P2P (DCUTR).

Unit tests were added, and the 3-phase MuSig2 example was updated to
utilize this new NAT-T connection logic. Tests confirm it is working,
but will be testing with real-world network configurations in the near
future.