Allow pubkey advertisement if signature valid
This commit changes the security logic to allow a new `peerId` to
advertise the same public key as another peer. This is ONLY allowed
because the signature of the advertisement has already been validated,
which confirms that the peer, with a new `peerId`, owns the private key.

It is up to the client implementation to generate a P2P keypair and use
the same private key at each startup, but not all peers will do this;
for a DHT server, it is important to accommodate these situations
without sacrificing security.